vphy: Fix bug selecting the wrong rx ref_clk
The rx refclock that the driver should use is selected with the device
tree property 'xlnx,rx-refclk-sel'. Petalinux generates a value of 0
in the devicetree when selecting 'refclk0' in Vivado and a value of 1
when selecting 'refclk1'.
According to ug576 table 2-10 the values to be programmed are:

0b001: GTREFCLK0 selected
0b010: GTREFCLK1 selected

so we have to add a 1 to the value from the devicetree before
programming to the 'Reference Clock Selection (RCS) Register (0x0010)'.